Always fork, because Ant misbehaves when calling the main class.
If the main method returns, Ant calls System.exit() and kills the
VM. This is almost always wrong.

Also, if an exception is thrown, many Ant layers are part of the
stacktrace making it hard to quickly assess the problem.

Finally, it's just an unsafe way to test an application since the
environment isn't clean.